fixed dead code, duplicated composable, unused districtId paramter, documentation
